Use the Post/Redirect/Get pattern instead.
This way the user will not be left with a POST as their last request.
When that's the case, if the user reloads the page, the browser will
prompt with a warning that it will resend the same form data. Instead,
we'll save the information in Flask's 'session' dictionary during the
first submittal, and use it after the redirect.