Add a change e-mail method.
This is used when the user returns with the correct token. Ensure that
the HMAC signature is correct, that the link clicked is for the actual
user that wants the e-mail changed, that the desired new e-mail address
isn't empty, and that the e-mail doesn't already exist in our database.